Adds discord support via the AttachmentsSlackFormatter.
