+++ src/Narochno.Serilog.Slack/SlackBatchingSink.cs
@@ -5,8 +5,6 @@ using Serilog.Events;
 using System;
 using Narochno.Slack;
 using Narochno.Serilog.Slack.Formatting;
-using Serilog.Debugging;
-using Narochno.Slack.Entities;
 
 namespace Narochno.Serilog.Slack
 {
@@ -24,11 +22,7 @@ namespace Narochno.Serilog.Slack
 
         protected override async Task EmitBatchAsync(IEnumerable<LogEvent> events)
         {
-            var result = await slackClient.PostMessage(messageFormatter.CreateMessage(events));
-            if (result != SlackCode.Ok)
-            {
-                SelfLog.WriteLine("Got error response from Slack: {0}", result);
-            }
+            await slackClient.PostMessage(messageFormatter.CreateMessage(events));
         }
 
         protected override void Dispose(bool disposing)
