+++ src/Narochno.Serilog.Slack/project.json
@@ -1,5 +1,5 @@
 {
-    "version": "1.2.0",
+    "version": "1.3.0",
     "authors": [ "alanedwardes", "adamhathcock" ],
 
     "packOptions": {
@@ -8,7 +8,7 @@
     },
 
     "dependencies": {
-        "Narochno.Slack": "1.6.0",
+        "Narochno.Slack": "2.0.0",
         "Serilog.Sinks.PeriodicBatching": "2.1.0"
     },
